Figured out and fixed the problem with crashes resulting from the dialog fragment for new users
Redirected calls from onStart() into the firebase calls, because the
calls were being executed before the data needed for those calls was
being retrieved. Also change how the dialog for selecting icons works,
there was no need to pass the user because the user was going to be null
anyway if it reached the condition to open up the dialog. Tested the
changes by deleting and adding new users and it seems to be working like
a charm now; doesn't crash/show multiple dialogs anymore