Cleaned up code for push notifications, found a bug and cant fix it
for some reason, if a user is set in Batch, they are permanently set for
the lifetime of the installation. This means that push notifications for
a user that was at one time logged in will still be sent to the phone,
even if that user had logged out. This may be intentional, I emailed
support to see what they say; I can't find any errors on the server so..